패키지명 변경 (Book -> book, Entity -> entity), adminGoodsController의 adminRepository 연동된 코드들 adminService로 이동
