--- src/main/java/Book/yes48/Entity/order/OrderStatus.java
@@ -1,6 +1,0 @@
-package Book.yes48.Entity.order;
-
-public enum OrderStatus {
-
-    ORDER, CANCEL
-}
