--- src/main/java/Book/yes48/Yes48Application.java
@@ -1,25 +1,0 @@
-package Book.yes48;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.web.config.PageableHandlerMethodArgumentResolverCustomizer;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class Yes48Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Yes48Application.class, args);
-	}
-
-	// Pageable Custom
-	@Bean
-	public PageableHandlerMethodArgumentResolverCustomizer customize() {
-		return p -> {
-			p.setOneIndexedParameters(true);	// 1부터 시작
-			p.setMaxPageSize(12);				// size=12
-		};
-	}
-}
