--- src/main/java/Book/yes48/config/QuerydslConfiguration.java
@@ -1,22 +1,0 @@
-package Book.yes48.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * 매번 사용하기 때문에 config로 작성
- */
-@Configuration
-public class QuerydslConfiguration {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(em);
-    }
-}
