--- src/main/java/Book/yes48/config/WebConfig.java
@@ -1,23 +1,0 @@
-package Book.yes48.config;
-
-import Book.yes48.web.controller.AdminController;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-
-    private String AdminConnectPath = "/admin/**";
-    private String GoodsConnectPath = "/book-list/**";
-    private String resourcePath = "file:///C:/upload/";
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler(AdminConnectPath, GoodsConnectPath)
-                .addResourceLocations(resourcePath);
-    }
-
-}
