--- src/main/java/Book/yes48/form/admin/search/AdminSearchCondition.java
@@ -1,41 +1,0 @@
-package Book.yes48.form.admin.search;
-
-import Book.yes48.Entity.FileStore;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class AdminSearchCondition {
-
-    private Long id;
-    private String name;
-    private String sort;
-    private String author;
-    private String publisher;
-    private String publisherDate;
-    private Integer price;
-    private Integer stockQuantity;
-    private String event;
-    private String state;
-    private FileStore fileStore;
-    private String searchBy;
-    private SearchType searchType;
-
-    public AdminSearchCondition(Long id, String name, String sort, String author, String publisher, String publisherDate, Integer price, Integer stockQuantity, String event,
-                                String state, FileStore fileStore, String searchBy, SearchType searchType) {
-        this.id = id;
-        this.name = name;
-        this.sort = sort;
-        this.author = author;
-        this.publisher = publisher;
-        this.publisherDate = publisherDate;
-        this.price = price;
-        this.stockQuantity = stockQuantity;
-        this.event = event;
-        this.state = state;
-        this.fileStore = fileStore;
-        this.searchBy = searchBy;
-        this.searchType = searchType;
-    }
-
-}
