--- src/main/java/Book/yes48/web/controller/HomeController.java
@@ -1,15 +1,0 @@
-package Book.yes48.web.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-
-}
