--- src/main/java/Book/yes48/web/controller/MemberController.java
@@ -1,13 +1,0 @@
-package Book.yes48.web.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MemberController {
-
-    @GetMapping("/join")
-    public String join() {
-        return "member/join";
-    }
-}
