+++ src/main/java/book/yes48/config/WebConfig.java
@@ -1,0 +1,22 @@
+package book.yes48.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig implements WebMvcConfigurer {
+
+    private String AdminConnectPath = "/admin/**";
+    private String GoodsConnectPath = "/book-list/**";
+    private String resourcePath = "file:///C:/upload/";
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler(AdminConnectPath, GoodsConnectPath)
+                .addResourceLocations(resourcePath);
+    }
+
+}
