+++ src/main/java/book/yes48/entity/order/OrderStatus.java
@@ -1,0 +1,6 @@
+package book.yes48.entity.order;
+
+public enum OrderStatus {
+
+    ORDER, CANCEL
+}
