+++ src/main/java/book/yes48/form/admin/search/AdminGoodsSearch.java
@@ -1,0 +1,15 @@
+package book.yes48.form.admin.search;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class AdminGoodsSearch {
+
+    private String searchBy;
+    private SearchType searchType;
+
+    public boolean isEmpty() {
+        return searchType == null || searchBy == null;
+    }
+}
