+++ src/main/java/book/yes48/web/controller/HomeController.java
@@ -1,0 +1,14 @@
+package book.yes48.web.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+
+}
